Fix checkbox not changing class of the item after drag and drop
