+++ src/index.js
@@ -1,9 +1,10 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './style.css';
 import 'bootstrap-icons/font/bootstrap-icons.css';
-import checkboxesEvent from './_checkboxFunctions.js';
 // eslint-disable-next-line import/no-named-default
-import { default as dragObjects, populate, list } from './_dragFunctions.js';
+import { default as checkboxesEvent, list } from './_checkboxFunctions.js';
+// eslint-disable-next-line import/no-named-default
+import { default as dragObjects, populate } from './_dragFunctions.js';
 
 populate();
 dragObjects();
